ao_coreaudio: Handle buffer underruns
We should set the buffer size to the number of samples we actually read.
It's not crystal clear from apple's docs whether this is allowed, but
I see examples of it on the web. On the other hand, console gives some
warning messages about frame mismatch and whatnot, but things seem
to work. There are reports that things don't work well (on newer
osx versions) if you have multiple audio output threads, but since
that never happens with mpv proper, this seems ok.

Backport 0341a6f1d39801160322d3fe16245f8387735f4b