Fix OSD PBO upload for arm mac
Honestly not sure what's going on. Tested a few things, and it seems
that when using a PBO buffer source, glTexSubImage2D() with a width
that doesn't match the stride of the underlying buffer causes issues.
The destination texture size doesn't seem to matter here, which makes
things even more confusing because nowhere is the stride of the
underlying PBO explicitly specified, so I don't know where this
dependency comes in. It's implicitly used to calculate the total
linear PBO size however. Since a partial height works fine, it's
only a partial width that causes issues, I'm guessing that somehow
when doing a glTexSubImage2D() with a PBO the only working operation
is uploading a continuous portion of the PBO.
This thus limits us to mapping and uploading the full-width,
but we can do partial heights.

(It doesn't seem to be an alignment issue since uploading each row
one at a time instead of in bulk doesn't work either. It's not a matter
of making the width a power of 2 since PBO works fine for the video
texture which is stuff like 1280. And I tried aligning the width to
next multiple of 32, but that didn't work either. So the only thing
Ican think of is that internally maybe there's some coalescing going on
which results in a broken assumption that the region uploaded is
continuous in memory. Maybe something related to DMA, who knows).