sub, demux: improve behavior with negative subtitle delay/muxed subs Backport of cf2b7a4
Basically when 126126e was committed, it broke track-switching since
subs are not reinited every frame. ec5cb76 tried to solve this like
upstream did, by completely reinit the sub on track switch. But this
actually introduces the issue that sub switches when paused won't show
internal subtitle tracks since those are lazily read and the packets
may not be ready.

It's not as simple as just reconfiguring the sub again on track switch
though, since that results in subs only showing up if the needed
packets are already demuxed. But if in a lazy-stream you switch to a
new region then try displaying subs, they won't appear immediately. This
is basically again a consequence of lazy streams, we need to explicitly
wait for those packets to be ready.

This is actually one of a set of related issues with lazy streams and
seeking: keyframe seek doesn't show subs,
sub-seek doesn't work properly, or this pause behavior "carrying over"
old subs. For all of these, when using --cache=yes things would work
because we already have the packets ready in that case.

Thus we introduce the workaround to allow force-fetching lazy streams.
This is a bit of a hack, but the scope is small enough that it should
not cause any issues.