Revert "player: don't cache subtitles across deselection"
This reverts commit ec5cb76b4dcfdd0dc5223942b449d76b476b7bdb.
The next commit will fix the main reason why this was needed to fix
broken track-switch behavior after the switch to selectively reconfigure
libass (instead of doing it every frame).

With that properly fixed, and with the lazy-stream-readahead-on-pause
which ensures that we not only have access to the last decoded packet
when toggling subs at a position but in fact demux the packet for
any seek position on demand (e.g. pause + [disable subs, seek, enable])
there's not too much benefit in having the ass events cached instead
of discarded and rebuilt anew on track switch.

But I guess there is still possibly one benefit to having the event
cache, namely you can use keyframe seeks to navigate and still
get subs (if the events list for that region was indeed already built).
It also doesn't really hurt to have this, and could possibly better
handle some weirdly muxed tracks that the lazy-stream-readahead mechanism
would bail out on.

Since as of yet we haven't backported the update_hard control event yet,
this does mean we lose that mechanism to "refresh" after runtime opt
changes. Whatever.