opengl: Fix additional corner-case for PBO with M1
We also need to check to make sure the pbo width is actually equal
to the texture width. Previously I had some confusion as to whether
we actually need the region-copy width to match the texture width
or the PBO width. It turns out that it has to be the PBO width, since
I ran into the example seen in the comments where PBO is larger than
texture width, but rendering still failed.

Also obviously we can never copy over more than the texture width
so this means that the two have to match exactly.