sub: Only force read on pause when track switching
Previously the force-read condition was set within update_subtitle
itself which meant it applied to all calls when paused, not just
track switching. But this doesn't make much sense, as the only
other thing you can do when paused is seeking. And it's the case that
for seeks we either already demux the necessary packets if it's an
exact seek, or for a keyframe seek we won't demux the packets even with
a force readahead since the relevant sub packet may be before the
keyframe.

Plus the whole point of keyframe seek is to be fast, while this
would only slow it down.