demux: Add missing in->seeking condition for lazy need wait
Last demux pts is only reset to npos after the seek is picked up
by the demux thread and finished. But if we issue a track switch
before that happens (e.g. demux thread has currently dropped
lock on fill buffer and is waiting for network, so has not serviced
the seek yet) then last demux pts we see would be wrong. Obviously we
wouldn't have any reader head set yet so we should ensure the client
retries.