sub: Make force read packet async when track switching
This is actually a much better fix than unilaterally disabling the
hard-fetch mechanism for networked streams. Essentially, the root issue
is that the sub packet (force)read from demuxer is
done in the call chain for when the track switch is issued. This means
that if things take too long, then we end up blocking the player core.
This was unnoticeable for on-disk files, but for slow/stalled network
streams, we might be waiting a long time.

To fix this, we should use non-blocking/async reads and move retries
into the playloop logic. This is actually not too bad, because there
already exists some code to deal with force-refreshing subs when no
video is currently playing, which we can build on top of.