player/osd: Don't update vo osd on every playloop iteration when paused
It turns out part of the reason why vo osd was originally updated
along with term osd was that it benefited from the debouncing.
Restore back this behavior. But we can actually go further and just
avoid the video.c updating osd entirely, so it's done only from the
playloop and we rely on explicit update flag for immediate response.