sub: Fix sub-update when no VO and paused
So it turns out that unlike mentioned in
92441362b83d4953fe3e726b3ca6a38b687f08d9
there is in fact one other case where we read sub packets when paused,
and that's where we seek without any video currently playing. In such
cases we need to explicitly call update_subtitles from the main
playloop ourselves, which the previous commit broke.

In fact the previous assumption of having an API that force-reads when update_subtitles
is called when paused probably wasn't a good idea since video.c would call it during seek anyhow.
So let's explicitly add this as a parameter.