vo: Use atomics to access some "read-only" properties
This probably will be reverted. It's not worth the tradeoff to clutter
the code, since these property accesses are mainly only used by stats
panel. Plus theoretically the acquire/release would slow things down
on the "hot path" of VO thread, whereas it's OK to make the playloop
wait a bit.