cocoa_common: Don't change window size on reinit if it's the same as old
The "legacy" cocoa backend handled this correctly, if we reinit but
the returned size is same as before we shouldn't do anything. This is
not the same as what the check currently did since if you reinit,
resize the window, then reinit again it would snap back to the old size
because it only compared against the current window frame
(which has changed).