cocoa-cb/libmpv: Improve thread safety
Basically the API guarantee libmpv provides is that we need to
ensure pretty much all of the _render functions are serialized
with respect to each other. The set_param calls were not obeying
that before this commit.

Additionally, we should make things easy on the API user by allowing
report_present/swap to "fail gracefully", if the context is already
uninit'ed. This is important since requiring the user to stop all
callbacks before the uninit finishes (or to serialize them against
uninit) would be tricky, and might lead to visual jarringness.