cocoa-cb: Avoid excessive copying of opts
Tbh I don't know the old actually copied or not since nothing in
swift makes sense, but we should only copy the opts struct
once during init and reconfig. Note that we need to make a copy
here since it's not safe to make use of it from a non-VO thread,
since it might be in the process of being modified by config_cache
updates.

(Technically the new situation of making a shallow-copy isn't quite
right either since it doesn't copy over strings, but it's better
than the old approach).