vo.c: remove opts callbacks
Seemed a bit weird for the VO to need to know about GPU-backend
specific stuff. Generalize the VOCTRL name and add corresponding
option field.

Backports
073858fcddddc757dcfb0e10640d19ac8d6d925d
531868fe0d2a35fbbff78d9a9ff8f96df73e69fd

Note that this actually goes back to the state in
d2bdb72b69536370c3046107fc593896e74803c3

In theory having the VO backend handle the opts themselves
(instead of requiring voctrl) would have been cleaner, but the effort
to do this was never completed. And as it stood, this resulted
in a weird mishmash of vo.c needing to know about gl-specific opts
and register change listener for them.