cocoa-cb: Use the resize voctrl for live-resize as well.
Previously we forced a redraw to get the async state updated. However
now that libmpv support resize voctrl, we may as well just use that.

We also use this opportunity to more strictly enforce that isAsync and
inLiveResize are always in sync. It was probably fine before, but basically
we never want to end up in a situation where libmpv calls our update callback
and we skip calling display because we think we're in a live-resize but
actually isAsync is set to false.