demux: Start cache search from keyframe_earliest
Previously we were using queue head. But now that we track the earliest
keyframe, we can start from there directly. We never return anything
that's not a keyframe, so all packets between head and that earliest
keyframe would be skipped anyway.

The pruning code pruned ensured that head always started with
the earliest keyframe anyhow though, and usually there won't be
many pre-keyframe packets so this is a minor optimization. But it
makes things conceptually cleaner.