demux: allow backward cache to use unused forward cache
The more significant effect of this is that it allows range-joining
to work more effectively by allowing the max_bw bytes worth of leeway
before pruning actually kicks in.

Before this, if we navigated to a new seek range we would see that
the backbuffer was immediately > cap, and start pruning that other
range aggressively, even though there's no need to do so.

Backports 5c0a626deec07614fd783316dc4679fbd0a66be8