playloop: Repeated backstep - avoid double-counting demxuer seek offset
Previously we used 1.1 * frame_dist as seek-back distance.
This is the recommendation for normal hr-seeks since it tries to get
the demuxer to demux slightly before the target pts (and the playloop
will decode until received frame >= target pts). However, the
SEEK_STRICT mode already ensures that we start from a keyframe
that is strictly able to reach our target-pts (i.e. our target frame is
always a trailing picture, higher in both DTS and PTS). As a result,
if our situation for consecutive frames is like

prev
!
|---*---|
        ^
        current

If we ask the demuxer to seek to the point marked (*) in backward
SEEK_STRICT mode, we are guaranteed to start from a kf that can reach the
immediately previous frame. For safety we also do the same epsilon
correction for the seek target, nudging it back by half a frame
so that we will terminate on our target prev and we're less likely
to overshoot the previous frame (it's not clear if such nudging is
actually required, but it doesn't hurt).

Note that most other video-players implement back-stepping purely via
hr-seeks in this fashion, just seeking 1/fps before (and letting
the decoder drop unnecessary frames before target.) mpv is a bit
unique in that it actually decodes up to the current frame
and displays the last frame before current pts.

It's not clear why things were done this way, it's certainly more robust
against broken files (and more likely to e.g. work with types with
incorrect pts, since it completely re-uses the forward playback logic.)