playloop: Fix offset calculation for delayed tracks
I'm pretty sure this was wrong. It added together the delay for sub
and audio, but that doesn't make sense since the demux offset
doesn't really compose in that fashion. Moreover the way it took
the rolling max over the cumulative sum is clearly wrong since
it means {10, -5} and {-5, 10} would give different results (on the first
the cumulative sum is always postiive, while not true for second.)