demux: Avoid repeated seek-pts adjustment for demuxer with sub-demuxer
If we used seek_strict and called a demuxer which itself had a sub-demuxer
(currently only demux_disc or demux_timeline) we could end up
applying the adjustment again. Avoid this by hackily using a flag
to indicate when adjustment has already been performed, and avoid
performing it again.

This mechanism should be used if hr-seek offset adjustment is ever
moved into the demuxer.