ao_coreaudio: register hotplug_cb in normal init() as well
Similar to 2e43bf19e1f260436a69ce33fac79439858e39fc but I have no idea
what that commit is trying to do with the refcout. The hotplug situation
with mpv is a bit weird, basically hotplug listeners are only used
for the purposes of populating the device-list property, and hotplug
listeners are only registered the first time this property is accessed.

Moreover, rather than re-using any exisitng AO context, a completely
new AO context is created. It's not exactly clear why this was done
(f061befb3338c5cd6837701b4232dad48340ab6c) but maybe it's for platforms
with multiple AO backends? Or probably it's if you don't currently
have audio playing but still want to observe the hotplug property,