cocoa-cb/window: Send resize event on full-screen enter
Full-screen is a combination of two things: our own custom animation
that changes the frame to the full-screen frame, and the system's
full-screen transition that actually makes it full-screen.

If the animation duration is zero, then we don't get any didEndLiveResize
for this custom animation, we only have the single resize event. This
was being erroneously filtered out.

Upstream mpv has the same issue.
See 4d6601924a4d2df557f78f3192236d84f7a4274e for a similar issue.