cocoa-cb: Redo double buffering approach
So it turns out that CAOpenGLLayer doesn't really support proper double buffering at all. What it has is more like page-flipping with a single front buffer. That is, if you draw, commit, then before that FBO (IOSurface) is displayed on screen draw to another FBO and try to commit that, the second one will replace the first. So a frame ends up being "eaten". The CATransaction flush/commit is what actually commits things to the screen, and is what must be carefully timed to avoid such a situation.

Our previous approach was something like

* render
* [transaction flush], pendingSwap++
* wait for pending swap if needed

For 2 pending swaps in flight this does not work since as mentioned above CAOpenGLLayer does not support true double buffering. With this commit we use the approach of:

* render, glFlush, pendingSwap++ & wait for pending swap if needed
* [transaction flush]

which ensures we don't do a transaction flush when something is already pending to be displayed.