cocoa-cb: Don't wait for swap during live-resize
1) Ensure that VO part waits for flush

Should really be part of the previous commit. For correctness the VO side needs to wait until
after the flush, as only then is the pending swap count correct for presentation tracking purposes.

We also take this opportunity to make "report flush" idempotent, in that you can call it several times
for a given render cycle without incrementing the total pending swap count more than once. This is basically
just for convenience so the application side doesn't need to add state to track whether or not it's already
been called for conditional codepaths.

2) For async mode (live resizing) we should not block the main thread

It's just a bad idea (tm) since the layer's async mode already does throttling for us. So we introduce a "skip" bit
in the call to wait for swap. We still increment the number of pending swaps, so we're really relying on the async throttling to avoid the number of pending swaps blowing up. Also add a failsafe that resets the pending swaps on async end just in case.

Note that it seems this introduces a possibility of a dropped frame when ending live resize. Not too sure why this is, it's probably something to do with the mismatch between the async mode timings (which are done off of a cftimer and not synchronized against vblank) and our regular timings (which are vsync aligned). Or when in async mode there is a queue that's not properly drained when switching back to sync mode. Whatever, I doubt most people would care.