cocoa-cb: Fix callback timing, use single buffer context
1) Switch to using a single-buffered opengl context. Since layer backed views draw to an FBO anyway and CAOpenGLLayer doesn't do any sort of double-buffering on the IOSurfaces, a single-buffered context is fine. And probably cheaper? Also with single-buffer we can avoid the presenting glFlush and just use flushRenderAPPLE which is an extension that does flush without present. (It actually seems that things work even if we don't explicitly flush. I couldn't find any flush calls anywhere, but presumably there has to be? Or maybe things are implicitly streamed instead of being queued up. Who knows...)

2) Based on my experimentation and analysis, the compositing happens in the first ~6% (the vblank) of the entire 16ms vsync interval. So this means that if you swap immediately upon the cvdisplaylink callback, the time it gives you is actually 1 frame too far ahead.

The more serious implication is that compositing could happen _anywhere_ within that 6%, so there isn't a set point we can divide the vsync interval into to say that if you flush before that point, and flush again after that point, they're guaranteed to both be visible. Instead what we have is an entire forbidden range, where we should ideally avoid any flushes within that entire 6% since they may possibly conflict with the previous or next flush.

Now usually since we will have 1 pending swap and we gate until after our callback, we can control the offset so that for almost all cases we do commit outside the forbidden period. But when we have no pending swaps and don't block, we could possibly commit anywhere (e.g. we have a 6% of chance of comitting within that bad region).

This is something to do for the future, now we just use 15% as the callback threshold instead of 50%, since 50% is definitely a bit too high (although it is the absolutely conservative option). 6% seems safe, but I'm using 15% instead just to be extra cautious. At one point I did try 10% and I think I saw a single drop frame out of like hundreds on the clock test but maybe it was my imagination and I'm going insane.

Note that previously before all this double-buffer refactoring (i.e. the state ~6 commits back) we ended up committing about 5% into the interval (validated both empirically
and it also makes sense conceptually since we synchronize at 0, then we know from past cocoa-gl experiments that rendering loop takes about 5%.)

Revisiting the comment there, it mentions that unblocking at 0.5 into the interval resulted in skipped frames. I think that was the case with the stats overlay
pulled up. It's known to be intensive, so adding half a vsync-interval delay time seems plausible. So if you started at the 1/2 mark, you would end up committing
inside the critical period.