cocoa-cb: Reduce critical vsync window to 10%
I found that the previous framedrops I thoguht I was seeing
only happen when you have two displays enabled. It's not clear what's
going on there: either the cvdisplaylink selected is wrong, or when
you have multiple displays the compositor does something different.
This needs more investigation, but for the single-display case 10%
should be fine.

To reproduce the issue with multiple displays:

1) have 2 displays, both with different referesh rate (I use promotion & 60fps)
2) Issue most noticeable right after start, or after a live-resize

Not sure what's going on here, and it happens too rarely to make it easy to debug.
It doesn't really matter since we can't fix it on our side anyhow:
1) Either it's a displaylink issue giving wrong timings, which doesn't seem likely since I at least verified the inteval is good
2) Maybe the window compositor does something different with multiple screens, it composites at a different time?
3) Maybe the window compositor simply takes longer than 15% of the interval with multiple screens.

I tried doing some probing but could not get it to reproduce again after I reset the mode on the display.
Maybe something for the future... the only real fix I can think of on our side either involves changing how we
create the display link if it's a display link issue (probably unlikely), or just changing the critical interval
to like 56% or so [which does appear to be the "safe" option].

Maybe https://thume.ca/2017/12/09/cvdisplaylink-doesnt-link-to-your-display/ is relevant,
I need to try looking at the phase.