DOCS: Update docs to reflect removed sub-scale-by-window
Also note in the docs that scale-with-window does not affect margins.
While we could theoretically make it do so, it wouldn't make any sense.

Basically the behavior from libass perspective is:
margins off: font scales with video size, margins respect to video size
margins on: font scales with window size, margins respect to window size

All scale-with-window does is force font to always scale with the
window-size, even in the margins off case. This can be useful in some
cases, if e.g. you want to zoom-out the video without affecting the
subtitle scale.

Now if margins are off but scale-with-window is true, the only really
sensible thing is to still have margins scale by the video size, because
they are necessarily interpreted with respect to the sides of the video.
I.e. if it didn't scale and the margins were always a fixed absolute
distance from e.g. the bottom of the video, then you would effectively
have things "shifting up" as you scaled the video down.