sub: Remove scale-to-window option.
Libass already does the sensible thing and scales
fonts with respect to the video size if ``--sub-use-margins`` is disabled
or with respect to the window-fit size if margins is enabled.
(The window-fit size is not necessarily the window size in the case aspect
ratio of video differs from that of the frame).