vo: Handle non-monotonic display time after presentation
In the case of double or triple-buffering when presentation feedback
is not used, using the time after present() as your display time
can result in bad timestamps that are not properly spaced apart.
This can cause the first few frames from a "fresh" present (empty
front buffer) to have incorrect timings, and as the swapchain
depth increases the actual a/v diff can also increase because the
timestamps used are increasingly further apart from the actual display
time.

Attempt to fix this by detecting such bad cases and overriding the
display time to be 1 vsync interval after the previous.