ra_gl: Use glBufferData for PBO if possible over bufferSubData
When using non-persistent buffers and we're not round-robin'ing,
we rely on the driver to detect a glBufferSubData that replaces
the entire buffer contents and convert it to orphan/splat behavior
that allocates new memory if needed instead of waiting for sync
of the previous buffer.

Not all drivers do this (apparently osx might not), so it's better
to be explicit and use glBufferData instead.

See https://stackoverflow.com/questions/18720983/plotting-real-time-data-with-opengl
https://hacksoflife.blogspot.com/2015/06/glmapbuffer-no-longer-cool.html
https://www.tumblr.com/thothonegan/135193767243/glbuffersubdata-vs-glbufferdata

If the old behavior is desired, explicitly setting num-pbo-buffers to 1
will force a single buffer to always be re-used.