coreaudio: Fix polling for physical format change
Previously the while loop was a bit questionable since it only
ever actually ran through twice making it a bit redundant. The comment
claims that it is not atomic so we may need to keep retrying even
after the listener was called. We may as well honor that.

Also add in something about why we use a semaphor here. This is something
that always puzzled me a bit. Well first since mpv reimplements its own
wrapper for osx semaphores (osx doesn't support posix unnamed semaphors
at all. But you do have gcd semaphores and mach semaphores). One good
reason to have semaphores is that they are async-signal/interrupt safe.
Unfortunately GCD semaphores are not, but mach semaphores are.

The mpv implementation reimplements semaphores using pipes which is also
signal safe (in fact it's explicitly called out as such).
But here that's not really relevant since we're not in a signal handler.
But I still argue it's better to use a semaphore as opposed to mutex+cvar
for three reasons:

1) It's generally better to avoid locks if possible when deadling with
   coreaudio for performance and predictability. Now this isn't a
   realtime audio thread, only a property callback, but then again
   the HAL notification thread is sort of an implementation detail
   that you shouldn't need to know about. (I think in practice
   it's just a dispatch queue).  However since in practice the
   implementation requires a syscall (same as mach semaphore),
   it can't really be argued to be any better than a GCD semaphore
   from this perspective. On osx pthread mutex has a fast-path
   as well which avoids kernel call so in practice this is a flimsy
   argument with current implementation.

2) The more important reason, is that it avoids temptation to get
   clever with the locking. In fact mutex/cvar used to be used but
   it quickly ran into deadlocks. See
   894000c434d060651e7cb533a942f06ced3937c0
   This was because the mutex was held too long, and coreaudio has
   internal mutexes so you could have a mutex ordering issue.
   If the mutex is not held for the duration of any coreaudio calls
   this should not be an issue, but just why risk it.