cocoa-gl: Fix fs glitch by delaying titlebar hiding
Unlike the cocoa-cb which has unified titlebar, cocoa-gl has an
explicit titlebar as part of the frame. It seems that adding/removing
the titlebar causes the layer to glitch.

Also doing it this way makes the transition to full-screen smoother
in that we no longer have the jarring resize at the very end. So
do the same technique for cocoa-cb as well.